requirements.txt
datasets
accelerate
transformers
trl
python-dotenv
wandb